science refactor
