+++ src/ros/rover/science/README.md
@@ -1,0 +1,7 @@
+# Science
+
+All ROS2 nodes and interfaces associated with our teams science functionality.
+
+See the science launch files for the latest configuration used in competition and testing:
+- [URC launch file](../nova_bringup/launch/urc_science.launch.py)
+- [ARC launch file](../nova_bringup/launch/arc_science.launch.py)
