+++ src/ros/rover/science/science/configs/commands.json
