+++ src/ros/rover/science/science/package.xml
