+++ src/ros/rover/science/science/science/archive/spectrometer_publisher.py
