+++ src/ros/rover/science/science/science/auger.py
