+++ src/ros/rover/science/science/science/pyCCDGUI-8apr19/pyCCDGUI/config.py
