+++ src/ros/rover/science/science/science/pyCCDGUI-8apr19/pyCCDGUI/pyCCDGUI.py
@@ -28,12 +28,9 @@
 
 #python imports
 import tkinter as tk
-import threading
 import queue
-import matplotlib as plt
 
 #application imports
-import config
 import CCDpanelsetup
 import CCDmenusetup
 import CCDplots
