+++ src/ros/rover/science/science/science/scimbal_cam.py
