+++ src/ros/rover/science/science/science/tests/test_distance.sh
