+++ src/ros/rover/science/science/setup.py
